// book_service/proto/bookPackage/CreateBookRequest.ts
// Original file: proto/book.proto


export interface CreateBookRequest {
  'name'?: (string);
  'author'?: (string);
  'publisher'?: (string);
  'addedBy'?: (string);
  'description'?: (string);
}

export interface CreateBookRequest__Output {
  'name'?: (string);
  'author'?: (string);
  'publisher'?: (string);
  'addedBy'?: (string);
  'description'?: (string);
}